requirements.txt
ultralytics
numpy
opencv-python
torch
torchvision
streamlit
pyyaml
pillow